Move Fineco tests into plugin folder
