+++ scripts/runJest.js
@@ -5,7 +5,7 @@ const os = require('os')
 const args = process.argv.slice(2)
 let cmd = 'jest --coverage'
 if (args[0] === 'fineco-it') {
-  cmd = 'jest fineco-it --runTestsByPath __tests__/fineco/e2e.sandbox.test.ts __tests__/fineco/smoke.test.ts'
+  cmd = 'jest fineco-it --runTestsByPath src/plugins/fineco-it/__tests__/e2e/e2e.sandbox.test.ts src/plugins/fineco-it/__tests__/e2e/smoke.test.ts'
   if (args.length > 1) cmd += ' ' + args.slice(1).join(' ')
 } else if (args.length > 0) {
   cmd += ' ' + args.join(' ')
