+++ src/plugins/fineco-it/__tests__/e2e/fixtures/credentials.json
